Always set repetition, not just when the transformation succeeds, so that the repetition does not unset itself when a repeat attempt fails
This trades one inconsistency for another: The initial mapping should not repeat when the transformation fails. We could achieve that for the line and visual mappings, although clumsily. (We need to pass a:isRepeat to the functions here; either we set up a duplicate set of Repeat <Plug>-mappings, or we try to detect the repetition at the beginning of the <Plug>-mapping (in s:Before(), comparing b:changedtick with g:repeat_tick, which caused funny errors when I tried it), and then pass this information to the later call via a script-local variable.) This cannot be achieved for the operator-pending mapping, because Vim will always repeat the g@ command; there's no way to undo that. Therefore, rather be internally consistent and avoid the large complexity of a partial solution (which other repeat.vim plugins also do not attempt.)