Add g:TextTransformContext.isAlgorithmRepeat
Add g:TextTransformContext.isRepeat. For that, we need to provide additional <Plug>TextR...
repeat mappings installed into repeat.vim. For the operator mapping, its g@ is repeated by Vim itself, but not the TextTransform#Arbitrary#Expression() triggered by the mapping, so we clear the repeatTick there to be able to distinguish.