Compatibility: Cannot pass :range to -count command after Vim 8.1.1241
We'd been using that trick to recognize a visual selection passed to the custom command vs. a passed [count]. That doesn't work any longer; we get "E14: Invalid address" now.
Specify -range instead of -count to allow passing ranges, and capture the (optional) count via -nargs=*. In order to distinguish :7Command from :Command 7, default the range to -1 and then either pass it or the arguments.